Remove useAirtableData and move WORKER_ACTIVITIES object to src/data/pages/index.js
